boom, I technically did it
This is interesting. It showcases how my brain tries to find shortcuts to avoid any meaningful commitment. I can try to compensate in the commit message, but it won't move the needle forward. But let me explain. These dummy commits are basically a way to tell myself 'do not worry about it tomorrow' when I have competing priorities and deadlines. It does not mean I won't do anything here later today, but I free myself to not worry. The day will be green today. A dummy commit is still a commit.