nannou: explorers and settles - initial setup
this is almost the same structure as the template, but I removed the squares animation. They should be static. Next step is to introduce some state to the, like I did in the particle_system_1 project. Each square is a cell that holds some state. We also need another structure for the agents that will fill the cells.